apps/oddo/requirements.txt
flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
sqlalchemy==2.0.21
flask-sqlalchemy==3.0.5
